// src/models/vehicle.model.ts
export interface VehicleModel {
    fuelTypeId: number;
    bodyId: number;
    colorId: number;
    userId: number;
    transmissionId: number;
    modelId: number;
    name: string;
    numberOfDoors: string;
    numberOfSeats: string;
    price: number;
    mileage: number;
    productionYear: number;
    horsepower: number;
    kilowatts: number;
}